Functionality for Clicking a Line
Created functionality for if a specific betting line is clicked. The background will turn to the secondary color and I will capture the "price" and "point" betting data. I accomplished all of this by creating state in this component that is made up of an array of 3 objects, representing the spread, total, and moneyline. This allowed me to create a "isClicked" key that could be toggled with a click event.