Added beslipRemoved State
When the betslip was visible as a sidebar (large screen width) and a slip was exited, this did not effect the color of the line that was clicked (and thus created that betslip). To solve the problem I took the event id that gets passed to function exitClicked and setBetslipRemoved to that  id. Then in the "boxTeam" component I have a useEffect occur whenver the betslipRemoved state changes and in this useEffect I setBettingLines to account for the removed slip.