Updated BetslipRemoved Usage
When a bet was confirmed, the lines green background color was not being immediately removed. I used "setBetslipRemoved" in the "betConfirm" component to handle this. Since multiple bets could be confirmed at once I had to make "betslipRemoved.id" an array of values. As a result I had to update the usage of "setBetslipRemoved" in the "betslip" component as well.