Update BoxTeam Logic for If No Bookmakers
Encountered a situation where there were no bookmakers for a given game and thus an empty string was returned for matchup.bookmakers. I fixed the situation by automatically applying "-1" value to the indexes, which will result in empty lines. However, if the string is not empty then the indexes will be assigned proper values.